+++ test/utils/load-env.js
@@ -1,0 +1,5 @@
+const dotenv = require('dotenv');
+
+const ENV = process.env.ENV || '.env-test-max-version';
+
+dotenv.config({ path: `${__dirname}/../../${ENV}` });