+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [Unreleased]
 ### Added
-- Field Types - Support `TIMESTAMP(n)`, `SMALLINT(n)`, `BIGINT(n)` and `BIT(1)` column types (MySQL)
+- Field Types - Support `TIMESTAMP(n)`, `SMALLINT(n)` and `BIGINT(n)` column types (MySQL)
 
 ## RELEASE 2.6.3 - 2019-11-04
 ### Fixed