+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - Field Types - Support `TIMESTAMP(n)`, `SMALLINT(n)` and `BIGINT(n)` column types (MySQL)
 - Command generate - Add comments to generated models.
+- Command Generate - Generate the forest collection files.
 
 ### Changed
 - Commands - Drop environment variables.