+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # Change Log
 
 ## [Unreleased]
+### Added
+- Field Types - Support `TIMESTAMP(n)`, `SMALLINT(n)` and `BIGINT(n)` column types (MySQL)
+
 ### Fixed
 - Command Generate - Generate MongoDB belongsTo.
+
+## RELEASE 2.6.3 - 2019-11-04
+### Fixed
+- Command Generate - Fix generation of project with a mongo database.
 
 ## RELEASE 2.6.2 - 2019-10-31
 ### Fixed